Refactor: moved StudentDegree methods to StudentDegreeController; Renamed StudentDegreeViews to StudentView
