+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/student/dto/StudentView.java
@@ -2,10 +2,10 @@ package ua.edu.chdtu.deanoffice.api.student.dto;
 
 import ua.edu.chdtu.deanoffice.api.general.GeneralView;
 
-public class StudentDegreeViews {
-    public interface DetailAndDegree {}
-    public interface SearchSimpleDegrees {}
-    public interface SimpleAndDegrees extends GeneralView.Named {}
+public class StudentView {
+    interface DetailAndDegree {}
+    interface SearchSimpleDegrees {}
+    interface SimpleAndDegrees extends GeneralView.Named {}
 
     public interface Simple extends SearchSimpleDegrees, SimpleAndDegrees {}
     public interface Detail extends Simple, DetailAndDegree {}
